fix(ip-api-change): изменила адрес запроса на api с https, оптимизировала количество запросов
