--- src/components/Services/ServiceCard/types.ts
@@ -1,16 +1,0 @@
-export interface GeoData {
-	status?: string;
-	country?: string;
-	countryCode?: string;
-	region?: string;
-	regionName?: string;
-	city?: string;
-	zip?: string;
-	lat?: number;
-	lon?: number;
-	timezone?: string;
-	isp?: string;
-	org?: string;
-	as?: string;
-	query?: string;
-}
