+++ src/components/Services/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import WebsiteCreationModal from '@/components/WebsiteCreationModal';
 import services from '@/constants/Services';
@@ -8,11 +8,34 @@ import services from '@/constants/Services';
 import ProjectDescriptionModal from '../ProjectDescriptionModal';
 import ServiceCard from './ServiceCard';
 import styles from './style.module.scss';
+import { GeoData } from './types';
 
 export default function Services() {
 	const [openCreationModal, setOpenCreationModal] = useState<boolean>(false);
 	const [isPrDescriptionModalOpen, setPrDescriptionModalOpen] = useState(false);
 	const [activeService, setActiveService] = useState('');
+
+	const [IPAddress, setIPAddress] = useState('');
+	const [geoInfo, setGeoInfo] = useState<GeoData>({});
+
+	useEffect(() => {
+		getVisitorIP();
+		getIPInfo();
+	}, []);
+
+	const getVisitorIP = async () => {
+		await fetch('https://api.ipify.org')
+			.then((response) => response.text())
+			.then((data) => setIPAddress(data))
+			.catch((error) => console.error(`Не смог получить IP: ${error}`));
+	};
+
+	const getIPInfo = async () => {
+		await fetch(`https://freeipapi.com/api/json/${IPAddress}`)
+			.then((response) => response.json())
+			.then((data) => setGeoInfo(data))
+			.catch((error) => console.log(`Не удалось получить геоданные: ${error}`));
+	};
 
 	const handlerAskBtn = (id: string) => () => {
 		setActiveService(id);
@@ -39,6 +62,7 @@ export default function Services() {
 						<ServiceCard
 							key={service.id}
 							service={service}
+							geo={geoInfo}
 							openCreationModal={handleOpenCreationModal}
 							handlerAskBtn={handlerAskBtn}
 						/>
