+++ src/components/Services/types.ts
@@ -1,0 +1,15 @@
+export interface GeoData {
+	ipVersion?: number;
+	ipAddress?: string;
+	latitude?: number;
+	longitude?: number;
+	countryName?: string;
+	countryCode?: string;
+	timeZone?: string;
+	zipCode?: string;
+	cityName?: string;
+	regionName?: string;
+	isProxy?: boolean;
+	continent?: string;
+	continentCode?: string;
+}
